integrations/mcp-rag/scripts/performance_baseline.py
#!/usr/bin/env python3
"""Performance baseline measurement for DeepAgents MCP-RAG integration.

This script establishes baseline performance metrics for all retrieval strategies
to validate against the <2s raw retrieval and <8s full research targets.
"""

import asyncio
import time
import statistics
from typing import Dict, List, Any
from pathlib import Path
import sys
import json

# Add src to path for imports
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

from deepagents_mcp_rag.retrievers.factory import RetrieverFactory
from deepagents_mcp_rag.config import get_settings


async def measure_retrieval_latency(strategy: str, query: str, iterations: int = 5) -> Dict[str, float]:
    """Measure retrieval latency for a specific strategy.

    Args:
        strategy: Retrieval strategy name
        query: Test query
        iterations: Number of test iterations

    Returns:
        Dictionary with timing statistics
    """
    print(f"Testing {strategy} strategy...")

    try:
        retriever = RetrieverFactory.create(strategy, k=5)
        latencies = []

        for i in range(iterations):
            start_time = time.time()

            try:
                # Note: This will likely fail without proper document corpus
                # but we can still measure the overhead
                documents = await retriever.retrieve(query)
                latency = (time.time() - start_time) * 1000  # Convert to ms
                latencies.append(latency)
                print(f"  Iteration {i+1}: {latency:.2f}ms ({len(documents)} docs)")

            except Exception as e:
                # Expected for empty document store
                latency = (time.time() - start_time) * 1000
                latencies.append(latency)
                print(f"  Iteration {i+1}: {latency:.2f}ms (error: {type(e).__name__})")

        return {
            "strategy": strategy,
            "mean_ms": statistics.mean(latencies),
            "median_ms": statistics.median(latencies),
            "min_ms": min(latencies),
            "max_ms": max(latencies),
            "std_ms": statistics.stdev(latencies) if len(latencies) > 1 else 0,
            "iterations": iterations,
            "query": query,
        }

    except Exception as e:
        print(f"  ❌ Failed to create {strategy} retriever: {e}")
        return {
            "strategy": strategy,
            "error": str(e),
            "query": query,
        }


async def test_infrastructure_connectivity() -> Dict[str, Any]:
    """Test connectivity to infrastructure services."""
    results = {}

    # Test Redis
    try:
        import redis
        r = redis.Redis.from_url("redis://localhost:6379")
        r.ping()
        results["redis"] = {"status": "✅ Connected", "latency_ms": "< 1"}
    except Exception as e:
        results["redis"] = {"status": "❌ Failed", "error": str(e)}

    # Test Qdrant
    try:
        import httpx
        response = httpx.get("http://localhost:6333/")
        if response.status_code == 200:
            results["qdrant"] = {"status": "✅ Connected", "version": response.json().get("version")}
        else:
            results["qdrant"] = {"status": "❌ Failed", "status_code": response.status_code}
    except Exception as e:
        results["qdrant"] = {"status": "❌ Failed", "error": str(e)}

    # Test PostgreSQL
    try:
        import psycopg2
        conn = psycopg2.connect("postgresql://postgres:postgres@localhost:5432/deepagents")
        conn.close()
        results["postgresql"] = {"status": "✅ Connected"}
    except Exception as e:
        results["postgresql"] = {"status": "❌ Failed", "error": str(e)}

    return results


async def measure_mcp_server_performance() -> Dict[str, Any]:
    """Measure MCP server startup and response performance."""
    print("Testing MCP server performance...")

    try:
        from deepagents_mcp_rag.mcp.server import create_mcp_server

        # Measure server creation time
        start_time = time.time()
        mcp_server = create_mcp_server()
        creation_time = (time.time() - start_time) * 1000

        # Get server info
        server_info = {
            "creation_time_ms": creation_time,
            "status": "✅ Created successfully",
        }

        # Test tool registration
        try:
            # This is a basic test of the server structure
            tools_count = len(getattr(mcp_server, '_tools', {}))
            resources_count = len(getattr(mcp_server, '_resources', {}))

            server_info.update({
                "tools_registered": tools_count,
                "resources_registered": resources_count,
            })
        except Exception as e:
            server_info["tool_info_error"] = str(e)

        return server_info

    except Exception as e:
        return {
            "status": "❌ Failed to create MCP server",
            "error": str(e)
        }


async def run_performance_baseline():
    """Run comprehensive performance baseline tests."""
    print("=" * 60)
    print("DeepAgents MCP-RAG Performance Baseline")
    print("=" * 60)

    # Load settings
    settings = get_settings()
    print(f"Configuration loaded: {settings.log_level} logging")
    print()

    # Test infrastructure connectivity
    print("1. Infrastructure Connectivity Test")
    print("-" * 40)
    infrastructure = await test_infrastructure_connectivity()
    for service, status in infrastructure.items():
        print(f"  {service}: {status['status']}")
    print()

    # Test MCP server performance
    print("2. MCP Server Performance Test")
    print("-" * 40)
    mcp_performance = await measure_mcp_server_performance()
    print(f"  Server creation: {mcp_performance.get('creation_time_ms', 'N/A'):.2f}ms")
    print(f"  Status: {mcp_performance.get('status', 'Unknown')}")
    print()

    # Test retrieval strategies
    print("3. Retrieval Strategy Performance Test")
    print("-" * 40)

    test_queries = [
        "What is machine learning?",
        "How do neural networks work?",
        "Explain transformers in AI",
    ]

    strategies = ["bm25", "vector", "parent_doc", "multi_query", "rerank", "ensemble"]
    results = []

    for strategy in strategies:
        for query in test_queries[:1]:  # Test with first query only for baseline
            result = await measure_retrieval_latency(strategy, query, iterations=3)
            results.append(result)

            if "error" not in result:
                mean_ms = result["mean_ms"]
                status = "✅" if mean_ms < 2000 else "⚠️" if mean_ms < 5000 else "❌"
                print(f"  {strategy}: {mean_ms:.2f}ms avg {status}")
            else:
                print(f"  {strategy}: ❌ {result['error']}")

    print()

    # Generate summary report
    print("4. Performance Summary")
    print("-" * 40)

    working_strategies = [r for r in results if "error" not in r]
    if working_strategies:
        avg_latencies = [r["mean_ms"] for r in working_strategies]
        overall_avg = statistics.mean(avg_latencies)

        print(f"  Working strategies: {len(working_strategies)}/{len(results)}")
        print(f"  Average latency: {overall_avg:.2f}ms")
        print(f"  Target compliance (<2000ms): {sum(1 for ms in avg_latencies if ms < 2000)}/{len(avg_latencies)}")
    else:
        print("  ⚠️  No strategies successfully completed (expected without document corpus)")

    # Save detailed results
    report = {
        "timestamp": time.time(),
        "infrastructure": infrastructure,
        "mcp_server": mcp_performance,
        "retrieval_results": results,
        "summary": {
            "working_strategies": len(working_strategies),
            "total_strategies": len(results),
            "target_compliance_2s": sum(1 for r in working_strategies if r["mean_ms"] < 2000),
        }
    }

    # Save to file
    output_file = Path(__file__).parent.parent / "performance_baseline_report.json"
    with open(output_file, "w") as f:
        json.dump(report, f, indent=2)

    print(f"\n📊 Detailed report saved to: {output_file}")
    print("\n🔧 Next steps:")
    print("   1. Ingest sample documents to test actual retrieval")
    print("   2. Configure API keys for LLM integration testing")
    print("   3. Start Phoenix tracing for observability")


if __name__ == "__main__":
    asyncio.run(run_performance_baseline())